requirements.txt
pip==20.3.3
paho-mqtt==1.5.1
psutil==5.8.0